Product entity refactored
