+++ src/main/java/com/kodilla/ecommercee/repository/ProductRepository.java
@@ -1,6 +1,6 @@
 package com.kodilla.ecommercee.repository;
 
-import com.kodilla.ecommercee.Product;
+import com.kodilla.ecommercee.domain.Product;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
